requirement.txt
asgiref==3.8.1
dj-svg==0.3.1
Django==5.1.4
django-appconf==1.0.6
django-cors-headers==4.6.0
django-image-cropping==1.7
djangorestframework==3.15.2
Markdown==3.7
pillow==11.0.0
sqlparse==0.5.3
